+++ packages/client/src/pages/MainPage/containers/StatementEditorBox/StatementEditorActantTable/StatementEditorActantTable.tsx
@@ -6,21 +6,20 @@ import {
   StyledTHead,
   StyledTh,
 } from "./StatementEditorActantTableStyles";
-import { StatementEditorActantTableRow } from "./StatementEditorActantTableRow/StatementEditorActantTableRow";
+import { StatementEditorActantTableRow } from "./StatementEditorActantTableRow";
 import { AttributesEditor } from "../../AttributesEditor/AttributesEditor";
 import { AttributeButtonGroup } from "../../AttributeButtonGroup/AttributeButtonGroup";
 
 import {
   IActant,
-  IResponseGeneric,
   IResponseStatement,
   IStatementActant,
   IStatementProp,
 } from "@shared/types";
-import { ActantSuggester, ActantTag, CertaintyToggle, ElvlToggle } from "../..";
-import { Button, ButtonGroup, Input, Loader } from "components";
+import { ActantSuggester, ActantTag } from "../..";
+import { Button } from "components";
 import { FaTrashAlt, FaUnlink, FaPlus } from "react-icons/fa";
-import { useMutation, UseMutationResult, useQueryClient } from "react-query";
+import { UseMutationResult } from "react-query";
 import { actantPositionDict } from "@shared/dictionaries";
 
 interface FilteredActantObject {
@@ -117,7 +116,7 @@ export const StatementEditorActantTable: React.FC<StatementEditorActantTable> = 
           return actant ? (
             <ActantTag
               actant={actant}
-              short={false}
+              // fullWidth
               button={
                 userCanEdit && (
                   <Button
