--- packages/client/src/pages/MainPage/containers/StatementEditorBox/StatementEditorActantTable/StatementEditorActantTableRow/StatementEditorActantTableRowStyles.tsx
@@ -1,24 +1,0 @@
-import styled from "styled-components";
-
-interface StyledTr {
-  isOdd?: boolean;
-  isSelected?: boolean;
-  opacity?: number;
-}
-export const StyledTr = styled.tr<StyledTr>`
-  opacity: ${({ opacity }) => (opacity ? opacity : 1)};
-  td:first-child {
-    padding-left: ${({ theme }) => theme.space[1]};
-    padding-right: ${({ theme }) => theme.space[2]};
-  }
-  td:not(:last-child) {
-    width: 1%;
-  }
-`;
-export const StyledTd = styled.td`
-  padding-top: ${({ theme }) => `${theme.space[1]}`};
-  padding-right: ${({ theme }) => `${theme.space[2]}`};
-  padding-bottom: ${({ theme }) => `${theme.space[1]}`};
-  padding-left: 0;
-  font-size: ${({ theme }) => theme.fontSize["sm"]};
-`;
