+++ packages/client/src/pages/MainPage/containers/StatementEditorBox/StatementEditorActantTable/StatementEditorActantTableStyles.tsx
@@ -4,6 +4,7 @@ export const StyledTable = styled.table`
   width: 100%;
   border-spacing: 0;
   border-collapse: collapse;
+  background-color: palevioletred;
 `;
 export const StyledTHead = styled.thead`
   font-size: ${({ theme }) => theme.fontSize["sm"]};
@@ -14,4 +15,27 @@ export const StyledTh = styled.th`
   color: ${({ theme }) => theme.color["info"]};
   font-weight: ${({ theme }) => theme.fontWeight["light"]};
   padding-bottom: ${({ theme }) => theme.space[1]};
+`;
+
+interface StyledTr {
+  isOdd?: boolean;
+  isSelected?: boolean;
+  opacity?: number;
+}
+export const StyledTr = styled.tr<StyledTr>`
+  opacity: ${({ opacity }) => (opacity ? opacity : 1)};
+  td:first-child {
+    padding-left: ${({ theme }) => theme.space[1]};
+    padding-right: ${({ theme }) => theme.space[2]};
+  }
+  td:not(:last-child) {
+    width: 1%;
+  }
+`;
+export const StyledTd = styled.td`
+  padding-top: ${({ theme }) => `${theme.space[1]}`};
+  padding-right: ${({ theme }) => `${theme.space[2]}`};
+  padding-bottom: ${({ theme }) => `${theme.space[1]}`};
+  padding-left: 0;
+  font-size: ${({ theme }) => theme.fontSize["sm"]};
 `;
